refactor: giga permission commit

+++ src/app/api/uploadthing/core.ts
@@ -4,7 +4,7 @@ import { db } from "@/server/db";
 import { getSession } from "@/server/utils";
 import { headers } from "next/headers";
 import { asset } from "@/server/db/schema";
-import { tryCatchAsync } from "@/lib/try-catch";
+import { tryCatch } from "@/lib/try-catch";
 
 const f = createUploadthing();
 
@@ -21,26 +21,26 @@ export const fileRouter = {
             return { user: session.user };
         })
         .onUploadComplete(async ({ metadata, file }) => {
-            const assetElement = await tryCatchAsync(() =>
+            const assetElement = await tryCatch(
                 db
                     .insert(asset)
                     .values({
+                        type: "profile-image",
                         uploadthingId: file.key,
                         url: file.ufsUrl,
                         uploadedBy: metadata.user.id,
                     })
                     .returning(),
-            )
-                .onError(async () => {
-                    const utapi = new UTApi();
-                    await utapi.deleteFiles(file.key);
-                })
-                .unwrap({
-                    expectation: "expectSingle",
-                    errorMessage: "Failed to save uploaded image to database",
-                });
+            );
 
-            return { assetId: assetElement.id };
+            if (!assetElement.success || assetElement.data.length !== 1) {
+                const utapi = new UTApi();
+                await utapi.deleteFiles(file.key);
+                assetElement.unwrap();
+                return;
+            }
+
+            return { assetId: assetElement.data[0]!.id };
         }),
 
     videoAudioUploader: f({
@@ -59,27 +59,26 @@ export const fileRouter = {
             return { user: session.user };
         })
         .onUploadComplete(async ({ metadata, file }) => {
-            const assetElement = await tryCatchAsync(() =>
+            const assetElement = await tryCatch(
                 db
                     .insert(asset)
                     .values({
+                        type: "recording",
                         uploadthingId: file.key,
                         url: file.ufsUrl,
                         uploadedBy: metadata.user.id,
                     })
                     .returning(),
-            )
-                .onError(async () => {
-                    const utapi = new UTApi();
-                    await utapi.deleteFiles(file.key);
-                })
-                .unwrap({
-                    expectation: "expectSingle",
-                    errorMessage:
-                        "Failed to save uploaded video/audio to database",
-                });
+            );
 
-            return { assetId: assetElement.id };
+            if (!assetElement.success || assetElement.data.length !== 1) {
+                const utapi = new UTApi();
+                await utapi.deleteFiles(file.key);
+                assetElement.unwrap();
+                return;
+            }
+
+            return { assetId: assetElement.data[0]!.id };
         }),
 } satisfies FileRouter;
